Module Tezos_benchmark

module Ast : sig ... end

module Base_samplers : sig ... end

module Benchmark : sig ... end

This module defines several signatures to create a benchmark. * The most generic is S; it provides a rather complete control on benchmark creation. * Simple and Simple_with_num can be used to ease benchmark creation, with simpler parameters. Registration functions in Registration will convert these simple interfaces back to S. * All the common parameters for these signatures are declared in Benchmark_base.

module Benchmark_helpers : sig ... end

module Builtin_benchmarks : sig ... end

Private module for internal benchmarks

module Builtin_models : sig ... end

module Codegen : sig ... end

module Config : sig ... end

This module implements a schema for a json file that contains the configuration for all the benchmarks registered in Snoop. This configuration file has a tree-like structure that follows the Namespace hierarchy of the benchmarks. Its root element is always ".".

module Costlang : sig ... end

module Crypto_samplers : sig ... end

Functions for sampling (pk,pkh,sk) triplets.

module Csv : sig ... end

Save and load tables as CSV files.

module Dep_graph : sig ... end

Models associated to benchmarks have free variables. Some of these variables are to be inferred from the empirical data, but some others must be provided by other models and correspond to _dependencies_ of a model upon the result of another one.

module Fixed_point_transform : sig ... end

Transform multiplications by constants in a costlang expression to fixed point arithmetic. Allows to make cost functions protocol-compatible.

module Free_variable : sig ... end

module Generator : sig ... end

module Inference : sig ... end

module Maths : sig ... end

module Measure : sig ... end

module Model : sig ... end

module Namespace : sig ... end

Namespace implements a type for the name of the benchmarks in Snoop It can be seen as a path of several names, similar to a file system

module Ocamlformat : sig ... end

module Override : sig ... end

module Plot : sig ... end

prbnmcn-gnuplot: a self-contained interface to Gnuplot.

module Pyinference : sig ... end

module Pyinit : sig ... end

module Pytools : sig ... end

module Registration : sig ... end

The Registration module contains functions regarding the registration of benchmarks, models, parameters and commands for the Snoop tool.

module Scikit_matrix : sig ... end

module Sparse_vec : sig ... end

module Structure_samplers : sig ... end