Module Base_samplers.Adversarial

Sampling of "adversarial" values in the sense that they exhibit the worst-case performance of the usual comparison functions.

These samplers generate pairs of a common prefix that has a size uniformly sampled in prefix_size, and a list of card distinct (with high-probability) elements sharing this common prefix.

raises [Invalid_argument]
if prefix_size.max < prefix_size.min or if prefix_size.min < 0 or of card <= 0.

val integers : prefix_size:range -> card:int -> (Z.t * Z.t list) sampler

val strings : prefix_size:range -> card:int -> (string * string list) sampler

val bytes : prefix_size:range -> card:int -> (bytes * bytes list) sampler