Module Tezos_benchmark.Benchmark_helpers

val load_json : string -> (Tezos_base.TzPervasives.Data_encoding.json, exn) Stdlib.result

Load a JSON file

val make_progress_printer : Stdlib.Format.formatter -> int -> string -> unit -> unit

make_progress_printer ppf total message returns a progress monitor printer with an internal counter starting from 1. Each call of the printer accumulates the counter and prints a progress monitor line like "Message 12/300".

val int_encoding : int Tezos_base.TzPervasives.Data_encoding.t

Data encoding of OCaml int, using Data-encoding.int64

val filename_of_local_model_name : string -> string

Replace '/' characters in a local model name by "__"