Module Tezos_benchmark.Codegen

type solution = {

map : float Free_variable.Map.t;

scores_list : ((string * Namespace.t) * Inference.scores) list;

}

val get_codegen_destination : Registration.model_info -> string option

get_codegen_destinations model_info will return the generated code destination in which given model is used

val solution_encoding : solution Tezos_base.TzPervasives.Data_encoding.t

val pp_solution : Stdlib.Format.formatter -> solution -> unit

val load_solution : string -> solution

Load a solution form a binary or JSON file

val save_solution : solution -> string -> unit

Save the given solution to binary and JSON files

val solution_to_csv : solution -> Csv.csv

type code

Generated code for a model

type module_

val codegen : Model.packed_model -> solution -> Costlang.transform -> Namespace.t -> code

Build the cost function codes of the specified model

val codegen_models : (Namespace.t * Registration.model_info) list -> solution -> Costlang.transform -> (string option * code) list

codegen_models models solution transform generates the cost function codes of multiple models.

val comment : string list -> code

Make a comment

val make_toplevel_module : code list -> module_

Make a toplevel module (an .ml file) from the given list of codes

val pp_code : Stdlib.Format.formatter -> code -> unit

val pp_module : Stdlib.Format.formatter -> module_ -> unit

val get_name_of_code : code -> string option

Get the function name of code if it has

module Parser : sig ... end