Module Tezos_benchmark.Crypto_samplers

Functions for sampling (pk,pkh,sk) triplets.

This module provides an implementation for finite pools of random (pk,pkh,sk) triplets. The pools grows to the target size as the user samples from it. When the target size is reached, sampling is performed in a round-robin fashion.

module type Param_S = sig ... end

module type Finite_key_pool_S = sig ... end

module Make_finite_key_pool (Arg : Param_S) : Finite_key_pool_S

Create a finite key pool.