Module Tezos_benchmark.Csv

Save and load tables as CSV files.

Note that this module does not support the full specification of CSV but covers only the specific format used in Snoop.

type csv = string list list

The type of table. A list of rows. Each row must have the same number of columns

val concat : ?check_disjoint_headers:bool -> csv -> csv -> csv

Horizontally concat tables. The Tables must have the same number of rows.

val export : filename:string -> ?separator:char -> ?linebreak:char -> csv -> unit

Save a table to a CSV file

val export_stdout : ?separator:char -> ?linebreak:char -> csv -> unit

Print a table as CSV to stdout

val import : filename:string -> ?separator:char -> unit -> csv

Load a CSV file

val append_columns : filename:string -> ?separator:char -> ?linebreak:char -> csv -> unit

Extend the CSV file with the given columns