Module Solver.Unsolved

type t

val build : Namespace.t -> fvs_unapplied:Free_variable.Set.t -> Free_variable.Set.t -> t

build name ~fvs_unapplied fvs_applied builds an unsolved node of a benchmark of name name. fvs_unapplied is the set of free variables of the models of the benchmark without workload application. fvs_applied is the set of free variables of the models of the benchmark applied with workloads.