Module type Generator.S

type config

Configuration of the benchmark generator

type workload

Workload corresponding to a benchmark run

val create_benchmarks : rng_state:Stdlib.Random.State.t -> bench_num:int -> config -> (unit -> workload benchmark) list

Creates a list of benchmarks, ready to be run. The bench_num option corresponds to the argument command-line specified by the user on command line, but it can be overriden by the config-specific settings. This is the case for instance when the benchmarks are performed on external artifacts. The benchmarks are thunked to prevent evaluating the workload until needed.