Module Plot.Histogram

Preparing histograms.

val hist : points:r1 Data.t -> ?color:Color.t -> ?bins:int -> ?binwidth:float -> ?legend:string -> unit -> r2 spec

hist ~points ?color ?bins ?binwidth ?legend creates an histogram out of points.

The optional argument ?color specifies the color of the histogram.

The optional argument ?bins specifies the number of bins to be used when constructing the histogram.

The optional argument ?binwidth specifies the width of each bin: the bigger, the coarser the histogram.

The optional argument ?legend specifies the legend attached to the histogram. If neither the ?bins or ?binwidth parameter is specified, this command will let ?bins default to 20 and will let ?binwidth be automatically computed by Gnuplot.