Module Plot.Scatter

Preparing scatter plots.

val points_2d : points:r2 Data.t -> ?style:Style.t -> ?legend:string -> ?error_bars:r2 Data.t -> unit -> r2 spec

points_2d ~points ~style ?legend () creates a 2d scatter plot using style to select color and point shape.

The optional argument ?style defaults to Style.default

The optional argument ?legend defaults to "".

The optional argument ?error_bars, if present, is a sequence of r2 elements such that each element (y1, y2) is displayed as a vertical error bar. The error bar associated to the element (x, y) then correspond to the segment starting at (x, y + y1) and ending at (x, y - y2).

raises Invalid_argument
if the length of error_bars, if present, is not equal to that of points.

val points_3d : points:r3 Data.t -> ?style:Style.t -> ?legend:string -> unit -> r3 spec

points_3d ~points ~style ?legend creates a 3d scatter plot. See points_2d for the meaning of other options.