Module Tezos_benchmark__Crypto_samplers

Functions for sampling (pk,pkh,sk) triplets.

This module provides an implementation for finite pools of random (pk,pkh,sk) triplets. The pools grows to the target size as the user samples from it. When the target size is reached, sampling is performed in a round-robin fashion.