Module Tezos_benchmark__Fixed_point_transform

type cast_mode =

| Ceil

| Floor

| Round

type options = {

precision : int;
Number of bits to consider when decomposing the mantissa

max_relative_error : float;
Percentage of admissible relative error when casting floats to ints

cast_mode : cast_mode;

inverse_scaling : int;
The constant prettification will consider 1/inverse_scaling digits to be not significant.

resolution : int;
Resolution of the grid using when prettifying constants.

}

type fp_error =

| Bad_fpclass of Stdlib.Float.fpclass

| Negative_or_zero_fp

type fixed_point_transform_error =

| Variable_not_found of Tezos_benchmark.Free_variable.t

| Model_not_found of string

| Imprecise_integer_cast of float * float

| Term_is_not_closed of Tezos_benchmark.Free_variable.t

exception Bad_floating_point_number of fp_error

exception Codegen_error of fixed_point_transform_error

val default_options : options

val pp_fixed_point_transform_error : Stdlib.Format.formatter -> fixed_point_transform_error -> unit

val cast_mode_encoding : cast_mode Tezos_base__TzPervasives.Data_encoding.encoding

val options_encoding : options Tezos_base__TzPervasives.Data_encoding.encoding

val log10 : int -> int

val pow : int -> int -> int

val snap_to_grid : inverse_scaling:int -> resolution:int -> int -> int

val int_of_float : cast_mode -> float -> int

val assert_fp_is_correct : float -> unit

val cast_safely_to_int : float -> cast_mode -> float -> int

module type Fixed_point_lang_sig = sig ... end

module Fixed_point_arithmetic : functor (Lang : Fixed_point_lang_sig) -> sig ... end

module Prettify_constants : functor (P : sig ... end) -> functor (X : Tezos_benchmark.Costlang.S) -> Tezos_benchmark.Costlang.S with type 'a repr = 'a X.repr and type size = X.size

module Convert_mult : functor (P : sig ... end) -> functor (X : Tezos_benchmark.Costlang.S) -> sig ... end

module Apply : functor (P : sig ... end) -> Tezos_benchmark.Costlang.Transform