Module Tezos_benchmark.Crypto_samplers

Functions for sampling (pk,pkh,sk) triplets.

This module provides an implementation for finite pools of random (pk,pkh,sk) triplets. The pools grows to the target size as the user samples from it. When the target size is reached, sampling is performed in a round-robin fashion.

module type Param_S = sig ... end

module type P_Finite_key_pool_S = sig ... end

module V0 : sig ... end

module V1 : sig ... end

module V_latest : sig ... end

include module type of V_latest

module type Finite_key_pool_S = P_Finite_key_pool_S with type public_key_hash := Tezos_crypto.Signature.V_latest.Public_key_hash.t and type public_key := Tezos_crypto.Signature.V_latest.Public_key.t and type secret_key := Tezos_crypto.Signature.V_latest.Secret_key.t

module Make_finite_key_pool (Arg : Param_S with type algo := Tezos_crypto.Signature.V_latest.algo) : Finite_key_pool_S

Create a finite key pool.