Module Tezos_benchmark.Fixed_point_transform

Transform multiplications by constants in a costlang expression to fixed point arithmetic. Allows to make cost functions protocol-compatible.

type cast_mode =

| Ceil

| Floor

| Round

Modes of casting of float to int

type options = {

precision : int;
(*
Number of bits to consider when decomposing the mantissa
*)

max_relative_error : float;
(*
Percentage of admissible relative error when casting floats to ints
*)

cast_mode : cast_mode;

inverse_scaling : int;
(*
The constant prettification will consider 1/inverse_scaling digits to be not significant.
*)

resolution : int;
(*
Resolution of the grid using when prettifying constants.
*)

}

Parameters for conversion to fixed point

val default_options : options

val options_encoding : options Tezos_base.TzPervasives.Data_encoding.t

module Apply (P : sig ... end) : Costlang.Transform