Module Tezos_benchmark.Generator

type 'aspect probe = {

apply : 'a. 'aspect -> (unit -> 'a) -> 'a;

aspects : unit -> 'aspect list;

get : 'aspect -> float list;

}

A probe implements an instrumented apply function. The implementation of apply is meant to record in a side-effecting way the result of a benchmark for the given closure. get allows to retrieve the results.

type 'workload benchmark =

| Calculated : {

workload : 'workload;

measure : unit -> float;

} -> 'workload benchmark
(*
Calculated provides its own measure function, unlike the other benchmarks where Measure module times their closures
*)

| Plain : {

workload : 'workload;

closure : unit -> unit;

} -> 'workload benchmark

| With_context : {

workload : 'workload;

closure : 'context -> unit;

with_context : 'a. ('context -> 'a) -> 'a;

} -> 'workload benchmark

| With_probe : {

workload : 'aspect -> 'workload;

probe : 'aspect probe;

closure : 'aspect probe -> unit;

} -> 'workload benchmark

The type of benchmarks. Measurements of benchmarks are directly calculated by measure or measured as the execution time of closure. Some benchmark requires to set-up/cleanup artifacts when being run, in that case use With_context with a proper implementation of the with_context function.

module type S = sig ... end