Module Tezos_benchmark.Model

type (_, _, _) arity =

| Zero_arity : ('elt, 'elt, unit) arity

| Succ_arity : ('elt, 'b, 'a) arity -> ('elt, 'elt -> 'b, int * 'a) arity

Specifying model arity and eliminator type

val arity_0 : ('a, 'a, unit) arity

val arity_1 : ('a, 'a -> 'a, int * unit) arity

val arity_2 : ('a, 'a -> 'a -> 'a, int * (int * unit)) arity

val arity_3 : ('a, 'a -> 'a -> 'a -> 'a, int * (int * (int * unit))) arity

type ('a, 'b) eq =

| Eq : ('a, 'a) eq

val elim_arities : elt m1 m2 a. ('elt, 'm1, 'a) arity -> ('elt, 'm2, 'a) arity -> ('m1, 'm2) eq

module type Model_impl = sig ... end

Models are strongly typed: Model_impl.arg_type exposes what a model expects on input. The relation between arg_type and model_type is encoded through a value of type arity.

module type Instantiated = sig ... end

type 'arg model = (module Model_impl with type arg_type = 'arg)

module type Applied = functor (X : Costlang.S) -> sig ... end

type applied = (module Applied)

type _ t =

| Packaged : {

conv : 'workload -> 'arg;

model : 'arg model;

} -> 'workload t

| Preapplied : {

model : 'workload -> applied;

} -> 'workload t

type for_codegen =

| For_codegen : _ t -> for_codegen

val apply_model : 'arg -> 'arg model -> applied

module Instantiate : functor (X : Costlang.S) -> functor (M : Model_impl) -> Instantiated with type 'a repr = 'a X.repr and type size = X.size and type arg_type = M.arg_type

val make : conv:('a -> 'b) -> model:'b model -> 'a t

val make_preapplied : model:('a -> applied) -> 'a t

val apply : 'a t -> 'a -> applied

val add_model : 'arg model -> 'arg model -> 'arg model

val precompose : a b. ('a -> 'b) -> 'b t -> 'a t

val unknown_const1 : const:Free_variable.t -> (module Model_impl with type arg_type = unit)

val unknown_const2 : const1:Free_variable.t -> const2:Free_variable.t -> (module Model_impl with type arg_type = unit)

val linear : coeff:Free_variable.t -> (module Model_impl with type arg_type = int * unit)

val affine : intercept:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * unit)

val affine_split_const : intercept1:Free_variable.t -> intercept2:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * unit)

val quadratic : coeff:Free_variable.t -> (module Model_impl with type arg_type = int * unit)

val nlogn : intercept:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * unit)

val linear_sum : intercept:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * (int * unit))

val linear_max : intercept:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * (int * unit))

val linear_min : intercept:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * (int * unit))

val linear_mul : intercept:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * (int * unit))

val bilinear : coeff1:Free_variable.t -> coeff2:Free_variable.t -> (module Model_impl with type arg_type = int * (int * unit))

val bilinear_affine : intercept:Free_variable.t -> coeff1:Free_variable.t -> coeff2:Free_variable.t -> (module Model_impl with type arg_type = int * (int * unit))

val nlogm : intercept:Free_variable.t -> coeff:Free_variable.t -> (module Model_impl with type arg_type = int * (int * unit))

val trilinear : coeff1:Free_variable.t -> coeff2:Free_variable.t -> coeff3:Free_variable.t -> (module Model_impl with type arg_type = int * (int * (int * unit)))

val breakdown : coeff1:Free_variable.t -> coeff2:Free_variable.t -> break:int -> (module Model_impl with type arg_type = int * unit)

A multi-affine model in two parts. The breakpoint break indicates the point at which the slope changes coefficient.

val breakdown2 : coeff1:Free_variable.t -> coeff2:Free_variable.t -> coeff3:Free_variable.t -> break1:int -> break2:int -> (module Model_impl with type arg_type = int * unit)

A multi-affine model in three parts, with breakpoints break1 and break2. Expects break1 <= break2

val breakdown2_const : coeff1:Free_variable.t -> coeff2:Free_variable.t -> coeff3:Free_variable.t -> const:Free_variable.t -> break1:int -> break2:int -> (module Model_impl with type arg_type = int * unit)

breakdown2 with a non-zero value at 0